python-rag-ingestion/pyproject.toml
[project]
name = "rhoai-rag-ingestion"
version = "0.1.0"
description = "RAG data ingestion pipeline for RHOAI AI feature sizing"
readme = "README.md"
requires-python = ">=3.9"
authors = [
    {name = "RHOAI Team"}
]
license = {text = "Apache-2.0"}
dependencies = [
    "llama-index>=0.10.0",
    "llama-index-readers-github",
    "llama-index-readers-web", 
    "llama-index-embeddings-openai",
    "llama-index-llms-openai",
    "pyyaml>=6.0",
    "python-dotenv>=1.0.0",
    "click>=8.0.0",
]

[project.scripts]
rhoai-rag = "rhoai_rag_ingestion.cli:cli"

[project.optional-dependencies]
all = [
    "llama-index-readers-slack",
    "llama-index-readers-notion",
    "llama-index-readers-confluence",
]
dev = [
    "pytest>=6.0",
    "black",
    "flake8",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["rhoai_rag_ingestion"]

[tool.uv]
dev-dependencies = []
